Revert "corrected byte to int"
This reverts commit 5b3771b7b312ba3e896fdaa4c0d9470a297fdf31.

+++ Github_Tutorial.ino
@@ -13,7 +13,6 @@
 
 void setup()
 {
-  
   Serial.begin(9600);
   Serial.println("Trimpot read example");
 
@@ -22,7 +21,7 @@ void setup()
 
 void loop() 
 {
-  int myValue = 0;
+  byte myValue = 0;
   myValue = analogRead(A0);
   
   Serial.print("The value is: ");
